chore: relocate icons
