+++ src/lib/components/icons/LinkIcon.svelte
