+++ src/lib/components/icons/MoonIcon.svelte
