+++ src/lib/components/icons/SunIcon.svelte
